edit and add link in modal window

+++ views/v_bookmarks_myLinks.php
@@ -93,62 +93,15 @@
 				</blockquote>
 				
 				<div class="text-right delete-btn">
-					<a data-toggle="modal" data-target="#myModalDelete" class="btn btn-danger btn-sm">Delete</a>
-					<a data-toggle="modal" data-target="#myModalEdit"  class="btn btn-default btn-sm">Edit</a>
+					<a  class="btn btn-danger btn-sm" href="/bookmarks/delete/<?=$bookmark['bookmark_id']?>" onclick="return confirm('Are you sure to delete this link?');">Delete</a>
+					<a class="btn btn-default btn-sm" data-toggle="modal" data-target="#myModalEdit" href='/bookmarks/edit_form/<?=$bookmark['bookmark_id']?>'>Edit</a>
 				</div>		
 			</div> <!-- / .feed-display -->
-			
-			<!-- Modal Delete
-			=========================================== -->
-			<div class="modal fade" id="myModalDelete" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
-				<div class="modal-dialog">
-					<div class="modal-content">
-						<div class="modal-body">
-							<button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
-							<p>Are you sure to delete this link?</p><br/>&nbsp;&nbsp;&nbsp;&nbsp;
-							<a class="btn btn-danger btn-sm" href="/bookmarks/delete/<?=$bookmark['bookmark_id']?>">Delete</a>
-							<a class="btn btn-default btn-sm" data-dismiss="modal" aria-hidden="true">Cancel</a>
-						</div>
-					</div><!-- /.modal-content -->
-				</div><!-- /.modal-dialog -->
-			</div><!-- /.modal -->
-		
-			<!-- Modal Edit
-			 =========================================== -->
-			<div class="modal fade" id="myModalEdit" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
-				<div class="modal-dialog">
-					<div class="modal-content">
-						<div class="modal-body">
-							<button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
-							<form method="post" action="/bookmarks/edit/<?=$bookmark['bookmark_id']?>" id="addLink">
-							
-								<!-- URL input -->
-								<label for="url">URL</label>
-								<input class="form-control" size="50" value="<?=$bookmark['url']?>" placeholder="url" id="url" type="text" name="url" required><br/>
-								<p id="urlError"></p>
-								
-								<!-- Title input -->
-								<label for="title">Title</label>
-								<input class="form-control" size="50" value="<?=$bookmark['title']?>" placeholder="title" id="title" type="text" name="title" required><br/>
-								<p id="titleError"></p>
-								
-								<!-- Comment Textarea -->
-								<label for='notes'>Comment</label><br>
-								<textarea class="form-control" rows="3" name='notes' id='notes' placeholder="What's your comment?" required><?=$bookmark['notes']?></textarea><br/>
-								<p id="notesError"></p>
-											 
-								<div class="text-right">
-									<button  type="submit" class="btn btn-success btn-sm">Update Link</button>&nbsp;&nbsp;<a class="btn btn-default btn-sm" data-dismiss="modal" aria-hidden="true">Cancel</a>
-								</div>
-							</form>
-						</div>
-					</div><!-- /.modal-content -->
-				</div><!-- /.modal-dialog -->
-			</div><!-- /.modal -->
 		<?php endforeach;?>
-		
 	</div> <!-- / .col-md-offset-2 .col-xs-12 .col-sm-10 .col-md-8 -->
 </div> <!-- / .row -->
+<div class="modal fade" id="myModalEdit" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
+</div><!-- /.modal -->
